fix: bump temperature to 16 bits
Latest firmware change these parameters to be 16 bit. We need to do some
work to allow for higher sized types when parsing canopen messages.

+++ src/Objects.hpp
@@ -16,8 +16,8 @@ namespace motors_roboteq_canopen {
     CANOPEN_DEFINE_OBJECT(0x210D, 1, VoltageInternal,               std::uint16_t);
     CANOPEN_DEFINE_OBJECT(0x210D, 2, VoltageBattery,                std::uint16_t);
     CANOPEN_DEFINE_OBJECT(0x210D, 3, Voltage5V,                     std::uint16_t);
-    CANOPEN_DEFINE_OBJECT(0x210F, 1, TemperatureMCU,                std::int8_t);
-    CANOPEN_DEFINE_OBJECT(0x210F, 2, TemperatureSensor0,            std::int8_t);
+    CANOPEN_DEFINE_OBJECT(0x210F, 1, TemperatureMCU,                std::int16_t);
+    CANOPEN_DEFINE_OBJECT(0x210F, 2, TemperatureSensor0,            std::int16_t);
 
     CANOPEN_DEFINE_OBJECT(0x2110, 1, Feedback,                      std::int16_t);
     CANOPEN_DEFINE_OBJECT(0x2111, 0, StatusFlagsRaw,                std::uint16_t);
